+++ src/main/java/com/victor/EventDrop/filedrops/FileDropUtils.java
@@ -31,8 +31,6 @@ public class FileDropUtils {
     private int roomFileCountThreshold;
 
 
-
-
     /**
      * Calculates the total size of all files in a specific room.
      *